Rename `kits` package to `walletappkit`
This is in preparation for a proposed move to a standalone module/JAR.

See: 
https://github.com/bitcoinj/bitcoinj/issues/3299#issuecomment-1745436911
