+++ core/src/main/java/org/bitcoinj/walletappkit/package-info.java
@@ -18,4 +18,4 @@
  * High level wrapper APIs around the bitcoinj building blocks. WalletAppKit is suitable for many different types of
  * apps that require an SPV wallet.
  */
-package org.bitcoinj.kits;
+package org.bitcoinj.walletappkit;
